+++ toolchain/base/int_store.h
@@ -10,16 +10,164 @@
 #include "llvm/ADT/APInt.h"
 #include "llvm/ADT/SmallVector.h"
 #include "toolchain/base/mem_usage.h"
-#include "toolchain/base/value_ids.h"
 #include "toolchain/base/value_store.h"
 #include "toolchain/base/yaml.h"
+#include "toolchain/base/index_base.h"
 
 namespace Carbon {
+
+// Valid IDs which are associated with tokens during lexing need to fit into a
+// compressed storage space, which may influence the specific formulation of the
+// ID. Note that there may still be IDs either not associated with tokens or
+// computed after lexing outside of this range.
+constexpr int TokenIdBits = 23;
 
 // Forward declare a testing peer so we can friend it.
 namespace Testing {
 struct IntStoreTestPeer;
 }  // namespace Testing
+
+// Corresponds to a canonicalized integer value. This is used both for integer
+// literal tokens, and integer values in SemIR. These always represent the
+// abstract mathematical value -- signed and regardless of the needed precision.
+//
+// Small values are internalized into the ID itself. Large values are
+// represented as an index into an array of `APInt`s with a canonicalized bit
+// width.
+class IntId : public Printable<IntId> {
+ public:
+  using ValueType = llvm::APInt;
+
+  static const IntId Invalid;
+
+  static auto MakeIndexOrInvalid(int index) -> IntId {
+    CARBON_DCHECK(index >= 0 && index <= InvalidIndex);
+    return IntId(ZeroIndexId - index);
+  }
+
+  static auto MakeFromTokenPayload(uint32_t payload) -> IntId {
+    // Token-associated IDs are signed `TokenIdBits` integers, so force the sign
+    // extension from that bit.
+    constexpr int Shift = 32 - TokenIdBits;
+    return IntId(static_cast<int32_t>(payload << Shift) >> Shift);
+  }
+
+  // Tries to make a signed 64-bit integer into an embedded value in the ID, and
+  // if unable to do that returns the `Invalid` ID.
+  static auto TryMakeValue(int64_t value) -> IntId {
+    if (MinValue <= value && value <= MaxValue) {
+      return IntId(value);
+    }
+
+    return Invalid;
+  }
+
+  // Tries to make a signed APInt into an embedded value in the ID, and if
+  // unable to do that returns the `Invalid` ID.
+  static auto TryMakeSignedValue(llvm::APInt value) -> IntId {
+    if (value.sge(MinValue) && value.sle(MaxValue)) {
+      return IntId(value.getSExtValue());
+    }
+
+    return Invalid;
+  }
+
+  // Tries to make an unsigned APInt into an embedded value in the ID, and if
+  // unable to do that returns the `Invalid` ID.
+  static auto TryMakeUnsignedValue(llvm::APInt value) -> IntId {
+    if (value.ule(MaxValue)) {
+      return IntId(value.getZExtValue());
+    }
+
+    return Invalid;
+  }
+
+  // Construct an ID from a raw 32-bit ID value.
+  static constexpr auto MakeRaw(int32_t raw_id) -> IntId { return IntId(raw_id); }
+
+  constexpr auto is_valid() const -> bool { return id_ != InvalidId; }
+  constexpr auto is_value() const -> bool { return id_ > ZeroIndexId; }
+  constexpr auto is_index() const -> bool { return id_ <= ZeroIndexId; }
+
+  constexpr auto AsValue() const -> int {
+    CARBON_DCHECK(is_value());
+    return id_;
+  }
+
+  constexpr auto AsIndex() const -> int {
+    CARBON_DCHECK(is_index());
+    return ZeroIndexId - id_;
+  }
+
+  constexpr auto AsRaw() const -> int32_t { return id_; }
+
+  auto Print(llvm::raw_ostream& out) const -> void {
+    out << "int [";
+    if (is_value()) {
+      out << "value: " << AsValue() << "]";
+    } else if (is_index()) {
+      out << "index: " << AsIndex() << "]";
+    } else {
+      CARBON_CHECK(!is_valid());
+      out << "invalid]";
+    }
+  }
+
+  friend constexpr auto operator==(IntId lhs, IntId rhs) -> bool {
+    return lhs.id_ == rhs.id_;
+  }
+  friend constexpr auto operator<=>(IntId lhs, IntId rhs)
+      -> std::strong_ordering {
+    return lhs.id_ <=> rhs.id_;
+  }
+
+ private:
+  // We need all the values from maximum to minimum and a healthy range of
+  // indices to all fit within the token ID bits.
+  //
+  // We represent this as a signed TokenIdBits-bit 2s compliment integer. The
+  // sign extension from TokenIdBits to a register size can be folded into the
+  // shift used to extract from compressed bitfield storage.
+  //
+  // We then divide the smallest 1/4th of the space to indices, and the larger
+  // 3/4ths to embedded values. For 23-bits total this still gives us 2 million
+  // unique integers larger than the embedded ones, which would be difficult to
+  // fill without exceeding the number of tokens we can lex (8 million). For
+  // non-token based integers, the indices can continue downward to the 32-bit
+  // signed integer minimum.
+  //
+  // Note that the invalid ID can't be used with a token. This is OK as we
+  // expect invalid tokens to be *error* tokens and not need to represent an
+  // invalid integer.
+  static constexpr int TokenIdBitsShift = 32 - TokenIdBits;
+  static constexpr int32_t MaxValue =
+      std::numeric_limits<int32_t>::max() >> TokenIdBitsShift;
+  static constexpr int32_t ZeroIndexId = std::numeric_limits<int32_t>::min() >>
+                                         (TokenIdBitsShift + 1);
+  static constexpr int32_t MinValue = ZeroIndexId + 1;
+  static constexpr int32_t InvalidId = std::numeric_limits<int32_t>::min();
+  static constexpr int32_t InvalidIndex = ZeroIndexId - InvalidId;
+
+  // Document the specific values of these constants to help visualize how the
+  // bit patterns map from the above computations.
+  //
+  // Each bit is either `T` for part of the token or `P` as part
+  // of the available payload that we use for the ID:
+  //
+  // clang-format off: visualizing bit positions
+  //
+  //                           0bTTTT'TTTT'TPPP'PPPP'PPPP'PPPP'PPPP'PPPP
+  static_assert(MaxValue    == 0b0000'0000'0011'1111'1111'1111'1111'1111);
+  static_assert(ZeroIndexId == 0b1111'1111'1110'0000'0000'0000'0000'0000);
+  static_assert(MinValue    == 0b1111'1111'1110'0000'0000'0000'0000'0001);
+  static_assert(InvalidId   == 0b1000'0000'0000'0000'0000'0000'0000'0000);
+  // clang-format on
+
+  constexpr explicit IntId(int32_t id) : id_(id) {}
+
+  int32_t id_;
+};
+constexpr IntId IntId::Invalid(IntId::InvalidId);
 
 // A canonicalizing value store with deep optimizations for integers.
 //
