+++ toolchain/sem_ir/BUILD
@@ -36,6 +36,7 @@ cc_library(
         "//common:check",
         "//common:ostream",
         "//toolchain/base:index_base",
+        "//toolchain/base:int_store",
         "//toolchain/base:shared_value_stores",
         "//toolchain/base:value_ids",
         "//toolchain/diagnostics:diagnostic_emitter",
@@ -76,6 +77,7 @@ cc_library(
         "//common:ostream",
         "//common:struct_reflection",
         "//toolchain/base:index_base",
+        "//toolchain/base:int_store",
         "//toolchain/base:value_store",
         "@llvm-project//llvm:Support",
     ],
@@ -129,6 +131,7 @@ cc_library(
         "//common:map",
         "//common:ostream",
         "//common:set",
+        "//toolchain/base:int_store",
         "//toolchain/base:kind_switch",
         "//toolchain/base:shared_value_stores",
         "//toolchain/base:value_ids",
