+++ toolchain/sem_ir/id_kind.h
@@ -7,6 +7,7 @@
 
 #include <algorithm>
 
+#include "toolchain/base/int_store.h"
 #include "toolchain/sem_ir/ids.h"
 
 namespace Carbon::SemIR {
