+++ toolchain/sem_ir/typed_insts.h
@@ -8,6 +8,7 @@
 #include "toolchain/parse/node_ids.h"
 #include "toolchain/sem_ir/builtin_inst_kind.h"
 #include "toolchain/sem_ir/ids.h"
+#include "toolchain/base/int_store.h"
 #include "toolchain/sem_ir/inst_kind.h"
 
 // Representations for specific kinds of instructions.
